requirements.txt
pika==1.3.1
zeep==4.2.1
lxml==4.9.2
python-dotenv==0.19.2
